+++ packages/view-main/src/modules/setting/init.ts
@@ -2,7 +2,6 @@ import { onConnected, onRelease } from '@/modules/app/shared'
 import { getEnvLocale, i18n } from '@/plugins/i18n'
 import { createUnsubscriptionSet } from '@/shared'
 
-import { appState } from '../app/store/state'
 import { getSetting, registerRemoteSettingAction, updateSetting } from './store/action'
 import { initSetting as overwriteSetting } from './store/commit'
 import { settingEvent } from './store/event'
@@ -43,11 +42,11 @@ export const initSetting = () => {
           if (keys.includes('common.isShowAnimation')) {
             document.body.classList.toggle('no-animation', !setting['common.isShowAnimation'])
           }
-          if (keys.includes('common.font') && !appState.isFullscreen) {
+          if (keys.includes('common.font')) {
             if (setting['common.font']) document.documentElement.style.fontFamily = setting['common.font']
             else document.documentElement.style.removeProperty('font-family')
           }
-          if (keys.includes('common.fontSize') && !appState.isFullscreen) {
+          if (keys.includes('common.fontSize')) {
             document.documentElement.style.fontSize = `${setting['common.fontSize']}px`
           }
         })